make Hobby usable again
delegate responsibilites back to Hobby.jsx component
